requirements.txt
psycopg2==2.4.6
Django==1.5
South==0.7.6
django-annoying==0.7.7
django-debug-toolbar==0.9.4
django-redis-sessions==0.3.1
sorl-thumbnail==11.12
python-dateutil==2.2